requirements.txt
Pillow==9.2.0
requests==2.25.1
streamlit==1.10.0
streamlit_lottie==0.0.3
